+++ clients/core/src/managementConsole/courseConfigurator/graphComponents/edges/DataEdgeProps.tsx
