+++ clients/core/src/managementConsole/courseConfigurator/graphComponents/edges/IconEdge.tsx
