+++ clients/core/src/managementConsole/courseConfigurator/graphComponents/edges/ParticipantEdgeProps.tsx
