+++ clients/core/src/managementConsole/courseConfigurator/graphComponents/phaseNode/PhaseNode.tsx
@@ -6,8 +6,8 @@ import { Input } from '@/components/ui/input'
 import { Pen, TriangleAlert, FileInput, FileOutput } from 'lucide-react'
 import { Badge } from '@/components/ui/badge'
 import { MetaDataBadges } from './components/MetaDataBadges'
-import { useCourseConfigurationState } from '../zustand/useCourseConfigurationStore'
-import { CoursePhaseWithPosition } from '../interfaces/coursePhaseWithPosition'
+import { useCourseConfigurationState } from '../../zustand/useCourseConfigurationStore'
+import { CoursePhaseWithPosition } from '../../interfaces/coursePhaseWithPosition'
 import { useParams } from 'react-router-dom'
 import { useCourseStore } from '@/zustand/useCourseStore'
 import { useAuthStore } from '@/zustand/useAuthStore'
