+++ clients/core/src/managementConsole/courseConfigurator/graphComponents/phaseNode/components/DeletePhaseButton.tsx
