+++ clients/core/src/managementConsole/courseConfigurator/graphComponents/phaseNode/components/MetaDataBadges.tsx
