+++ clients/core/src/managementConsole/courseConfigurator/graphComponents/phaseNode/components/utils/getBadgeStatus.tsx
