+++ clients/core/src/managementConsole/courseConfigurator/graphComponents/phaseNode/components/utils/renderBadgeTooltip.tsx
