+++ clients/core/src/managementConsole/courseConfigurator/graphComponents/phaseNode/components/utils/renderTooltipList.tsx
