+++ clients/core/src/managementConsole/courseConfigurator/handlers/useComputeLayoutedElements.tsx
@@ -1,7 +1,7 @@
 import { useCourseConfigurationState } from '../zustand/useCourseConfigurationStore'
 import { getLayoutedElements } from '../utils/getLayoutedElements'
-import { ParticipantEdgeProps } from '../Edges/ParticipantEdgeProps'
-import { DataEdgeProps } from '../Edges/DataEdgeProps'
+import { ParticipantEdgeProps } from '../graphComponents/edges/ParticipantEdgeProps'
+import { DataEdgeProps } from '../graphComponents/edges/DataEdgeProps'
 
 export const useComputeLayoutedElements = () => {
   const { coursePhases, coursePhaseGraph, metaDataGraph } = useCourseConfigurationState()
